refactor: differ between graphing in terminal and in VS code
